backend/ml/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy>=1.21.0
opencv-python>=4.7.0
pillow>=9.0.0
tqdm>=4.65.0
scikit-learn>=1.0.0
wandb>=0.15.0
decord>=0.6.0 